Use SCAN and add support for redis://:password@hostname:port?db=1 urls
Use REDIS SCAN instead of KEYS
==

SCAN preferred to KEYS for production use, thus syncing use:
- http://redis.io/commands/scan
- http://redis.io/commands/keys

```
Warning: consider KEYS as a command that should only be used in
production environments with extreme care. It may ruin performance when
it is executed against large databases. This command is intended for
debugging and special operations, such as changing your keyspace layout.
Don't use KEYS in your regular application code. If you're looking for a
way to find keys in a subset of your keyspace, consider using SCAN or
sets.
```

Optional URL Input Format
==
Adds support for redis:// type urls which are used in redis-cli, with
the optional extension of `?db=1` if db needs to be non-zero.

Adds specs to verify the behavior. Preferentially chooses that type of
link parsing if the url starts with redis:// prefix. Otherwise falls
through to existing behavior.